aggregate.py: add tab format
Note that empty strings ('') and numbers should not be
on the same column, for tabulate interprets the whole
column as a string, thereby ignoring the floatfmt option.

Here we add tests to make sure that None instead of ''
is used with tabulate when we want some empty fields.
We also test the inverse when emitting CSV.